Cleanup client project Use spaces 4 to be consistent with the rest of the projects
