+++ src/VoltProjects.Client/.eslintrc.json
@@ -19,7 +19,7 @@
     "rules": {
         "indent": [
             "error",
-            "tab"
+            4
         ],
         "linebreak-style": [
             "error",
