+++ src/VoltProjects.Client/src/main.ts
@@ -1,16 +1,12 @@
 import { getPreferredTheme, setTheme } from './theme';
-import { renderAside } from './scripts/nav';
-import { renderToc } from './scripts/toc';
+import { renderAside } from './nav';
+import { initToc } from './toc';
 import { initSearchNav } from './searchNav';
-
-//Force include Bootstrap icon fonts
-import 'bootstrap-icons/font/fonts/bootstrap-icons.woff';
-import 'bootstrap-icons/font/fonts/bootstrap-icons.woff2';
 
 //Main styling
 import './scss/main.scss';
 
-//Bootstrap collapse, used by navbar, and as such every page
+//Bootstrap JS stuff we use
 import 'bootstrap/js/dist/collapse';
 import 'bootstrap/js/dist/dropdown';
 
@@ -19,10 +15,10 @@ declare global {
 }
 
 window.vGlobal = {
-	setTheme: setTheme
+    setTheme: setTheme
 };
 
 setTheme(getPreferredTheme());
 renderAside();
-renderToc();
+initToc();
 initSearchNav();
