--- src/VoltProjects.Client/src/scss/_variable.scss
@@ -1,37 +1,0 @@
-$border-radius: .375rem !default;
-
-//Set colors
-$primary: #1574a3;
-$dark: #121212;
-
-//Customize dark theme to look a bit better
-$body-color-dark: white;
-$body-bg-dark: $dark;
-$body-secondary-bg-dark: $dark;
-
-//Breaking Points
-$aside-break: 1400px;
-$grid-breakpoints: (
-    xs: 0,
-    sm: 580px,
-    md: 760px,
-    lg: 1100px,
-    xl: 1400px,
-    xxl: 1900px
-) !default;
-$container-max-widths: (
-    sm: 580px,
-    md: 760px,
-    lg: 1150px,
-    xl: 1400px,
-    xxl: 1800px
-) !default;
-
-//Make link decoration default have no decoration
-$link-decoration: none !default;
-$link-hover-decoration: underline !default;
-$link-color: $primary;
-//$link-color-dark: $primary;
-
-//Button color
-$btn-color: white;
