--- src/VoltProjects.Client/src/scss/core/alerts.scss
@@ -1,32 +1,0 @@
-@import "icons-font";
-
-.alert h5 {
-    text-transform: uppercase;
-    font-weight: 700;
-    font-size: 1rem;
-}
-
-.alert h5::before {
-    font-family: $bootstrap-icons-font;
-    position: relative;
-    margin-right: .5em;
-    top: .2em;
-    font-size: 1.25em;
-    font-weight: 400;
-}
-
-.alert-info h5::before {
-    content: map-get($bootstrap-icons-map, "info-circle-fill");
-}
-
-.alert-warning h5::before {
-    content: map-get($bootstrap-icons-map, "exclamation-circle-fill");
-}
-
-.alert-danger h5::before {
-    content: map-get($bootstrap-icons-map, "x-circle-fill");
-}
-
-.alert-success h5::before {
-    content: map-get($bootstrap-icons-map, "lightbulb-fill");
-}
