--- src/VoltProjects.Client/src/scss/core/buttons.scss
@@ -1,9 +1,0 @@
-.btn-outline {
-    --bs-btn-color: var(--bs-body-color);
-    --bs-btn-bg: var(--bs-body-bg);
-    --bs-btn-border-color: var(--bs-body-bg);
-    --bs-btn-hover-color: var(--bs-body-bg);
-    --bs-btn-hover-bg: var(--bs-body-color);
-    --bs-btn-hover-border-color: var(--bs-body-color);
-    --bs-btn-focus-shadow-rgb: 54, 54, 54;
-}
