--- src/VoltProjects.Client/src/scss/core/footer.scss
@@ -1,5 +1,0 @@
-.footer-container {
-  padding: 2rem 0;
-  margin-top: auto;
-  color: var(--bs-body-color) !important;
-}
