--- src/VoltProjects.Client/src/scss/core/hero.scss
@@ -1,28 +1,0 @@
-.hero {
-    .hero-content {
-        display: flex;
-        flex-direction: column;
-        padding: 5em 0;
-        gap: 1em;
-
-        h1 {
-            font-size: calc(1.525rem + 3.3vw);
-            font-weight: 600;
-        }
-
-        .lead {
-            font-weight: 400;
-            font-size: calc(1.275rem + .3vw);
-        }
-
-        .actions {
-            display: flex;
-            gap: 1em;
-
-            .btn {
-                font-weight: 600;
-                @include button-size(.75em, 1.5em, 1.25em, .3em);
-            }
-        }
-    }
-}
