--- src/VoltProjects.Client/src/scss/core/icons-font.scss
@@ -1,24 +1,0 @@
-//Based off bootstrap-icons's scss
-$bootstrap-icons-font: "bootstrap-icons" !default;
-$bootstrap-icons-font-src: url("#{$bootstrap-icons-font}.woff2") format("woff2"), url("#{$bootstrap-icons-font}.woff") format("woff") !default;
-
-$bootstrap-icons-map: (
-    "discord": "\f300",
-    "github": "\f3ed",
-    "envelope-fill": "\f32c",
-    "globe2": "\f3ef",
-    "twitter": "\f5ef",
-    "funnel": "\f3e1",
-    "info-circle-fill": "\f430",
-    "exclamation-circle-fill": "\f332",
-    "lightbulb-fill": "\f468",
-    "moon-fill": "\f494",
-    "pencil-fill": "\f4c9",
-    "brightness-high-fill": "\f1d1",
-    "search": "\f52a",
-    "three-dots-vertical": "\f5d3",
-    "printer-fill": "\f500",
-    "x-circle-fill": "\f622",
-    "mastodon": "\f647",  
-    "twitter-x": "\f8db",
-) !default;
