--- src/VoltProjects.Client/src/scss/core/nav.scss
@@ -1,3 +1,0 @@
-.navbar-brand {
-    font-weight: 600;
-}
