+++ src/VoltProjects.Client/src/scss/projects/builders/docfx.scss
