+++ src/VoltProjects.Client/src/scss/projects/layout.scss
@@ -1,3 +1,6 @@
+//
+// Main Project Layout
+//
 .layout {
     display: grid;
     margin-top: 1em;
@@ -72,7 +75,9 @@
     }
 }
 
+//
 // Metabar
+//
 .metabar {
     color: #999;
     font-size: 14px;
@@ -87,7 +92,6 @@
     margin: 0 6px;
 }
   
-// Metabar Contributors List
 .contributors {
     list-style: none;
     display: flex;
@@ -95,10 +99,58 @@
     padding: 0;
   
     & > li {
-      margin: 0 4px;
+        margin: 0 4px;
     }
   
     img {
-      border-radius: 50%;
+        border-radius: 50%;
     }
+}
+
+//
+// Aside
+//
+.in-this-article {
+    padding-top: .5em;
+    font-size: 14px;
+  
+    h6 {
+        font-size: 1.4em !important;
+    }
+  
+    h6::before {
+        content: "\f479";
+        font-size: 1.1em;
+        padding-right: 0.2em;
+    }
+  
+    .nav {
+        display: block;
+        padding-left: 16px;
+        margin-left: 4px;
+        border-left: 2px solid var(--bs-body-color);
+  
+        a {
+            padding: 5px 5px 0px 5px;
+        }
+
+        a.active {
+            border-left: 3px solid $primary;
+        }
+
+        a:hover {
+            text-decoration: underline;
+        }
+    }
+}
+  
+//
+// Search
+//
+.project-search {
+    border-color: $secondary;
+}
+
+.popover-search {
+    --bs-popover-max-width: 35em;
 }
