--- src/VoltProjects.Client/src/scss/projectview/_main.scss
@@ -1,10 +1,0 @@
-@import "layout";
-@import "nav";
-@import "toc";
-@import "syntaxHighlight";
-@import "projectSearch";
-
-@import "docfx";
-
-// Ideally we would get rid of the dotnet/docfx specific stuff
-@import "dotnet";
