--- src/VoltProjects.Client/src/scss/projectview/nav.scss
@@ -1,33 +1,0 @@
-.in-this-article {
-  padding-top: .5em;
-  font-size: 14px;
-
-  h6 {
-    font-size: 1.4em !important;
-  }
-
-  h6::before {
-    content: "\f479";
-    font-size: 1.1em;
-    padding-right: 0.2em;
-  }
-
-  .nav {
-    display: block;
-    padding-left: 16px;
-    margin-left: 4px;
-    border-left: 2px solid var(--bs-body-color);
-
-    a {
-      padding: 5px 5px 0px 5px;
-    }
-
-    a.active {
-      border-left: 3px solid $primary;
-    }
-
-    a:hover {
-      text-decoration: underline;
-    }
-  }
-}
