--- src/VoltProjects.Client/src/scss/projectview/projectSearch.scss
@@ -1,3 +1,0 @@
-.project-search {
-    border-color: $secondary;
-}
