+++ src/VoltProjects.Client/src/scss/setup.scss
@@ -1,0 +1,65 @@
+$font-family-sans-serif: 'Roboto', sans-serif;
+$font-family-monospace: 'Fira Code', monospace;
+
+//Other setup
+$border-radius: .375rem !default;
+
+//Set colors
+$primary: #1574a3;
+$dark: #121212;
+
+//Customize dark theme to look a bit better
+$body-color-dark: white;
+$body-bg-dark: $dark;
+$body-secondary-bg-dark: $dark;
+
+//Breaking Points
+$aside-break: 1400px;
+$grid-breakpoints: (
+    xs: 0,
+    sm: 580px,
+    md: 760px,
+    lg: 1100px,
+    xl: 1400px,
+    xxl: 1900px
+) !default;
+$container-max-widths: (
+    sm: 580px,
+    md: 760px,
+    lg: 1150px,
+    xl: 1400px,
+    xxl: 1800px
+) !default;
+
+//Make link decoration default have no decoration
+$link-decoration: none !default;
+$link-hover-decoration: underline !default;
+$link-color: $primary;
+//$link-color-dark: $primary;
+
+//Button color
+$btn-color: white;
+
+//Bootstrap icons
+$bootstrap-icons-font-src: url('bootstrap-icons/font/fonts/bootstrap-icons.woff2') format('woff2'), url('bootstrap-icons/font/fonts/bootstrap-icons.woff') format('woff');
+
+$bootstrap-icons-map: (
+    "discord": "\f300",
+    "github": "\f3ed",
+    "envelope-fill": "\f32c",
+    "globe2": "\f3ef",
+    "twitter": "\f5ef",
+    "funnel": "\f3e1",
+    "info-circle-fill": "\f430",
+    "exclamation-circle-fill": "\f332",
+    "lightbulb-fill": "\f468",
+    "moon-fill": "\f494",
+    "pencil-fill": "\f4c9",
+    "brightness-high-fill": "\f1d1",
+    "search": "\f52a",
+    "three-dots-vertical": "\f5d3",
+    "printer-fill": "\f500",
+    "x-circle-fill": "\f622",
+    "mastodon": "\f647",  
+    "twitter-x": "\f8db",
+) !default;
